pyproject.toml
[tool.poetry]
name = "ddns-headscale"
version = "0.1.0"
description = ""
authors = ["alblue <hslixdu@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
pyyaml = "^6.0.1"
pydnspod = "^0.0.2"
dns-dnspod = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
